requirements.txt
kaggle==1.6.17
python-dotenv==1.0.1